MS-80 Исправлена ошибка при удалении сеансов пользователя

+++ api/src/Service/UserSession/DeleteUserSessionService.php
@@ -4,6 +4,8 @@ namespace App\Service\UserSession;
 
 use App\Entity\User;
 use App\Entity\UserSession;
+use App\Enum\UserSessionTypeEnum;
+use App\Rest\Http\Exceptions\EntityNotFoundException;
 use App\Service\AbstractService;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
@@ -23,6 +25,10 @@ class DeleteUserSessionService extends AbstractService
      */
     public function make(UserSession $userSession): JsonResponse
     {
+        if ($userSession->getType() !== UserSessionTypeEnum::TEMP->name) {
+            throw EntityNotFoundException::userSession();
+        }
+
         $this->em->remove($userSession);
         $this->em->flush();
 
@@ -37,8 +43,12 @@ class DeleteUserSessionService extends AbstractService
     public function deleteAll(User $toUser): JsonResponse
     {
         $userSessionRepository = $this->em->getRepository(UserSession::class);
+        $userSessions = $userSessionRepository->findBy([
+            'user' => $toUser,
+            'type' => UserSessionTypeEnum::TEMP->name
+        ]);
 
-        foreach ($userSessionRepository->findBy(['user' => $toUser]) as $userSession) {
+        foreach ($userSessions as $userSession) {
             $this->em->remove($userSession);
         }
